Replace pre-ordering of justification with lazy one
so that it decides which justifications should be displayed first not
right after computing them but at the moment of displaying. It is
reasonable in case of large amount of justifications in total since user
definitely does not want to see them all but has to wait while they are
sorting anyway.