Reconfigure (justification) preferences structure
so that "Justifications" tab at "Explanation" Protege preferences
section is a section as well now. All the settings are "General" tab in
that section. There is a list of justification explanation plugins at
the tab also. Any justification plugin is able to use corresponding
justification preferences extension point in case it has its own
preferences tab to add.